% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.R
\name{align_dfcol}
\alias{align_dfcol}
\title{Align a data.frame to a template}
\usage{
align_dfcol(target, template)
}
\arguments{
\item{target}{data.frame}

\item{template}{data.frame}
}
\description{
align_dfcol
}
\examples{
target <- data.frame(matrix(NA, ncol = 5))
template <- data.frame(matrix(NA, ncol = 6))
names(target) <- c(letters[c(3, 1, 2, 4, 8)])
names(template) <- c(letters[1:6])
align_dfcol(target =  target, template = template)
}

