% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{break_word}
\alias{break_word}
\title{break_word}
\usage{
break_word(x, max.len)
}
\arguments{
\item{x}{character string}

\item{max.len}{maximum length of the first line}
}
\description{
Break a character string into two lines based on a specified maximum length
}
\examples{
break_word("a super package", 3)
break_word("a super package", 9)
}
