% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond2sal.R
\name{cond2sal}
\alias{cond2sal}
\title{Conductivity to salinity conversion}
\usage{
cond2sal(c, t = 25, P = 0)
}
\arguments{
\item{c}{numeric conducitivity in uS (measurements in mS need to multiplied by 1000)}

\item{t}{numeric celcius temperature}

\item{P}{numeric optional pressure (defaults to 0)}
}
\description{
cond2sal
}
\details{
algorithm based off an excel implementation by N. Iricanin
}
\examples{
cond2sal(c=52*1000)
}
\seealso{
\code{\link[wql]{ec2pss}}
}
