% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{Read a config.py file in R}
\usage{
config(file = "config.py")
}
\arguments{
\item{file}{a file path}
}
\description{
Read a config.py file in R
}
\examples{
\dontrun{
 config()
}
}
