% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{dd2dms}
\alias{dd2dms}
\title{dd2dms}
\usage{
dd2dms(x)
}
\arguments{
\item{x}{numeric vector of length 1 corresonding to decimal degrees}
}
\description{
Convert decimal degree coordinates to degrees, minutes, and seconds.
}
\examples{
dd <- 25.31015
dd2dms(dd)

dd <- -80.37198
dd2dms(dd)
}

