% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo.R
\name{download_zenodo}
\alias{download_zenodo}
\title{Download files from Zenodo record id}
\usage{
download_zenodo(record_id, dest_folder = "",
  token = Sys.getenv("ZENODO_PAT"))
}
\arguments{
\item{record_id}{record id}

\item{dest_folder}{destination folder for file downloads}

\item{token}{API key}
}
\description{
Download files from Zenodo record id
}
\details{
https://github.com/zenodo/zenodo/issues/1629#issuecomment-435062462
}
\examples{
\dontrun{
download_zenodo(record_id = "2025865")#' 
}
}
