% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_if_not_exists}
\alias{get_if_not_exists}
\title{Download a file if it doesn't already exist}
\usage{
get_if_not_exists(x, destfile, overwrite = FALSE)
}
\arguments{
\item{x}{url string or function to be evaluated}

\item{destfile}{file.path}

\item{overwrite}{logical force overwrite}
}
\description{
Download a file if it doesn't already exist
}
\examples{
\dontrun{
get_if_not_exists("http://www.omegahat.net/RCurl/data.gz", "data.gz")
junk_rds <- function(destfile){
     saveRDS(1, destfile)
}
get_if_not_exists(junk_rds, "junk.rds")
}
}
