% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{get_intersects}
\alias{get_intersects}
\title{Get intersecting sf subset}
\usage{
get_intersects(target, src)
}
\arguments{
\item{target}{target sf object}

\item{src}{source sf object}
}
\description{
Get intersecting sf subset
}
