% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{get_within}
\alias{get_within}
\title{Get sf objects within a source object}
\usage{
get_within(dt, bb)
}
\arguments{
\item{dt}{source sf object}

\item{bb}{bounding object}
}
\description{
Get sf objects within a source object
}
