% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gg_quantdot}
\alias{gg_quantdot}
\title{Create a quantile dotplot in ggplot2}
\usage{
gg_quantdot(dt, grp, var)
}
\arguments{
\item{dt}{data.frame}

\item{grp}{grouping variable}

\item{var}{dependent variable}
}
\description{
Create a quantile dotplot in ggplot2
}
\examples{
\dontrun{
gg_quantdot(mtcars, cyl, "mpg")
}
}
