% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{is_a_lake_gage}
\alias{is_a_lake_gage}
\title{Detect if a USGS gage is upstream of a lake}
\usage{
is_a_lake_gage(site_no, distance_threshold = 20)
}
\arguments{
\item{site_no}{character to preserve leading zeros}

\item{distance_threshold}{numeric in units of km}
}
\description{
Detect if a USGS gage is upstream of a lake
}
\examples{
\dontrun{
site_no <- "040871473"
is_lake_gage(site_no)
site_no <- "05078470"
site_no <- "05427718"

}
}
