% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{jheatmap}
\alias{jheatmap}
\title{jheatmap}
\usage{
jheatmap(dt, focal_columns = NULL)
}
\arguments{
\item{dt}{data.frame}

\item{focal_columns}{character vector specifying column names to focus on}
}
\description{
jheatmap
}
\examples{
\dontrun{
data(iris)
dt <- iris
jheatmap(dt)
jheatmap(dt, c("Sepal.Length"))
}
}
