% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsta-package.R
\docType{package}
\name{jsta-package}
\alias{jsta}
\alias{jsta-package}
\title{jsta: Jemma Stachelek's Personal Library of Functions}
\description{
A library of functions to manipulate date fields, calculate salinity from conductivity, pad string fields, calculate polar coordinates, ...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jsta/rjsta}
  \item Report bugs at \url{https://github.com/jsta/rjsta/issues}
}

}
\keyword{internal}
