% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{jstapply}
\alias{jstapply}
\title{A progress bar version of lapply without boilerplate overhead}
\usage{
jstapply(X, FUN, flag, ...)
}
\arguments{
\item{X}{a vector (atomic or list) or an \code{\link[base]{expression}}
    object.  Other objects (including classed objects) will be coerced
    by \code{base::\link[base]{as.list}}.}

\item{FUN}{the function to be applied to each element of \code{X}:
    see \sQuote{Details}.  In the case of functions like
    \code{+}, \code{\%*\%}, the function name must be backquoted or quoted.}

\item{flag}{placeholder not currently implemented to specify pre-bar flag}

\item{...}{optional arguments to \code{FUN}.}
}
\description{
A progress bar version of lapply without boilerplate overhead
}
\examples{
\dontrun{#' 
lapply(1:3, function(x) Sys.sleep(x))
# compared with:
jstapply(1:3, function(x) Sys.sleep(x))
}
}
