% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_wiki.R
\name{lake_wiki}
\alias{lake_wiki}
\title{lake_wiki}
\usage{
lake_wiki(lake_name, map = FALSE, ...)
}
\arguments{
\item{lake_name}{character}

\item{map}{logical produce map of lake location?}

\item{...}{arguments passed to maps::map}
}
\description{
lake_wiki
}
\examples{
lake_wiki("Lake Mendota")
lake_wiki("Lake Mendota", map = TRUE, "usa")
lake_wiki("Lake Nipigon", map = TRUE, regions = "Canada")
}

