% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{make_valid_geom_s2}
\alias{make_valid_geom_s2}
\title{Rebuild an sf object geometry to fix cross edges}
\usage{
make_valid_geom_s2(sf_object)
}
\arguments{
\item{sf_object}{An sf object to fix}
}
\description{
Rebuild an sf object geometry to fix cross edges
}
