% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_wiki.R
\name{map_lake_wiki}
\alias{map_lake_wiki}
\title{map_lake_wiki}
\usage{
map_lake_wiki(lake_name, ...)
}
\arguments{
\item{lake_name}{character}

\item{...}{arguments passed to maps::map}
}
\description{
map_lake_wiki
}
\examples{
\dontrun{
map_lake_wiki("Corey Lake", "usa")

map_lake_wiki("Lake Nipigon", regions = "Canada")
}
}

