% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{pdf_table}
\alias{pdf_table}
\title{pdf_table}
\usage{
pdf_table(x, out_name = "test.pdf")
}
\arguments{
\item{x}{data.frame or kable output}

\item{out_name}{character file name}
}
\description{
Create a pdf snapshot of a data.frame or kable output
}
\examples{
\dontrun{
data(iris)
pdf_table(iris[1:5,])
pdf_table(knitr::kable(iris[1:5,]))
}
}
