% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_invisible}
\alias{plot_invisible}
\title{plot_invisible}
\usage{
plot_invisible(...)
}
\arguments{
\item{...}{character plotting code to be evaluated}
}
\description{
Wrapper to silently create a plot
}
\details{
Stores plots as image files in the temporary workspace (See <http://stackoverflow.com/questions/20363266/how-can-i-suppress-the-creation-of-a-plot-while-calling-a-function-in-r>)
}
\examples{
plot_invisible("plot(1)")
}

