% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{point_in_poly}
\alias{point_in_poly}
\title{Return points in spatial (sp package) polygons}
\usage{
point_in_poly(dt, poly_shape)
}
\arguments{
\item{dt}{SpatialPointsDataFrame}

\item{poly_shape}{SpatialPolygonsDataFrame}
}
\description{
spatial join modified from spatialEco::point.in.poly
}
