% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar.R
\name{pol2cart}
\alias{pol2cart}
\title{Convert polar coordinates to cartographic (x,y) coordinates}
\usage{
pol2cart(theta, rho)
}
\arguments{
\item{theta}{numeric angle relative to the positive x-axis}

\item{rho}{numeric distance from the origin (0,0)}
}
\description{
pol2cart
}
\examples{
pol2cart(1.0443,1.9901)
}

