% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tabular}
\alias{tabular}
\title{Format a table for inclusion in roxygen docs}
\usage{
tabular(df, ...)
}
\arguments{
\item{df}{data.frame}

\item{\dots}{additional options passed to \code{format}}
}
\description{
Format a table for inclusion in roxygen docs
}
\examples{
df <- data.frame(a_b = 1, b_c = 2, stringsAsFactors = FALSE)
tabular(df)
}
\references{
<http://r-pkgs.had.co.nz/man.html>
}
