% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wygen.R
\name{wygen}
\alias{wygen}
\title{Generate Water Year(WY) tags based on a May 1st - April 30th WY.}
\usage{
wygen(dates)
}
\arguments{
\item{dates}{character date that passes coercion to POSIX formatting}
}
\examples{
dates<-"2014-01-01"
wygen(dates)
}

