% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{zero_pad}
\alias{zero_pad}
\title{zero_pad}
\usage{
zero_pad(x, digits)
}
\arguments{
\item{x}{character input to pad}

\item{digits}{numeric. number of zeros to pre-append}
}
\description{
zero_pad
}
\examples{
\dontrun{
zero_pad(1, 2)
zero_pad(1, 3)
}
}
