% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar.R
\name{cart2pol}
\alias{cart2pol}
\title{Convert cartographic (x,y) coordinates to polar coordinates}
\usage{
cart2pol(x, y)
}
\arguments{
\item{x}{numeric coordinate}

\item{y}{numeric coordinate}
}
\examples{
cart2pol(1,sin(6)+2)
}

