% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dates.R
\name{date456posix}
\alias{date456posix}
\title{Convert numeric dates in mddyy to POSIXct}
\usage{
date456posix(x, century)
}
\arguments{
\item{x}{numeric where the first 1-2 digits specify the month attribute because leading zeros have been stripped}

\item{century}{numeric century recommended choice of "19" or "20"}
}
\description{
Convert numeric dates in mddyy to POSIXct
}
\examples{
dates <- c("51514", "101214", "8714", "1214", "81412")
date456posix(dates, century = "20")
}

