% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_if_not_exists}
\alias{get_if_not_exists}
\title{Download a file if it doesn't already exist}
\usage{
get_if_not_exists(url, destfile, overwrite = FALSE)
}
\arguments{
\item{url}{url string}

\item{destfile}{file.path}

\item{overwrite}{logical force overwrite}
}
\description{
Download a file if it doesn't already exist
}
