% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gitignore}
\alias{gitignore}
\title{gitignore}
\usage{
gitignore(f, dry.run = FALSE)
}
\arguments{
\item{f}{file.path or wildcard to add to gitignore}
}
\description{
gitignore
}
\examples{
\dontrun{
gitignore("test")
}
}
