% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{key_state}
\alias{key_state}
\title{key_state}
\usage{
key_state(x)
}
\arguments{
\item{x}{object with an unabbreviated state column}
}
\description{
key_state
}
\examples{
key_state(data.frame(state.name = datasets::state.name, stringsAsFactors = FALSE))
}
