% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupkey.R
\name{lookupkey}
\alias{lookupkey}
\title{Convert character vector based on lookup values}
\usage{
lookupkey(key, input)
}
\arguments{
\item{key}{data.frame or matrix where the first column are the strings to be corrected and the second is the "corrected" definitions.}

\item{input}{character vector}
}
\description{
lookupkey
}
\examples{
key <- cbind(c("one", "two", "three", "four",
          "five"), c("one1", "two2", "three3", "over", "over"))
input <- c("one", "one", "two", "three", "four", "five")
lookupkey(key, input)

}
