% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_invisible}
\alias{plot_invisible}
\title{Invisibly push a base plot to a png file}
\usage{
plot_invisible(text, outpath)
}
\arguments{
\item{text}{character plotting code to be evaluated}

\item{outpath}{file.path}
}
\description{
Wrapper to silently create a plot
}
\details{
Stores plots as image files in the temporary workspace (See <http://stackoverflow.com/questions/20363266/how-can-i-suppress-the-creation-of-a-plot-while-calling-a-function-in-r>)
}
\examples{
\dontrun{
plot_invisible("plot(1)", "test.png")
}
}
