% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{svg_to_pdf}
\alias{svg_to_pdf}
\title{Convert an svg file to pdf}
\usage{
svg_to_pdf(url, outname = gsub(".svg", ".pdf", basename(url)))
}
\arguments{
\item{url}{url}

\item{outname}{outname}
}
\description{
Convert an svg file to pdf
}
\examples{
\dontrun{
svg_to_pdf(
"https://img.shields.io/badge/Paper-10.1002/lno.11137-blue.svg")
}
}
