% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{yyyymm_expand}
\alias{yyyymm_expand}
\title{Expand numeric dates in yyyymm to yyyy-mm}
\usage{
yyyymm_expand(x)
}
\arguments{
\item{x}{numeric in yyyymm format}
}
\examples{
dates <- c("200812", "201201")
sapply(dates, yyyymm2posix)
}

